Organized files
Made input_output files and added the correct functions inside
